+++ CHANGELOG.md
@@ -28,6 +28,7 @@ Please only add new entries below the [Unreleased](#unreleased---releasedate) he
 ### Features
 
 - **macros**: add `asset!` macros for asset management.(#798 @M-Adoo)
+- **macros**: add `include_asset!` macro for compile-time asset embedding. (#799 @M-Adoo)
 
 ### Breaking